In order to construct an object on demand, we can create a constructor function. It is creating an object, thus the name of the funciton should represents the object, not the operation. Also the naming convention is that it starts with capital letter. In order to use the constructor function, you need to prepend it with ‘new’ t.ex new Student(){ }
And it starts with function keyword. The reaosn why you cannot use the arrow function for the constructor function is that arrow function dont get access to ‘this’ in current context but they go level up. Therefore it would not refer to the object that it created by the function, but the parent context.

Function is a special js object. As a resul, function can have property just like any other js object type. This is what makes it possible to use ‘.prototype’ on constructor function, in order to add the method to the object.